Rename and refactor front components
