+++ client/src/components/Buttons.js
@@ -22,17 +22,3 @@ LinkButton.propTypes = {
   text: PropTypes.string.isRequired,
   className: PropTypes.string
 }
-
-export const ActionButton = ({ actionHandler, text, className }) => {
-  return (
-    <Button onClick={actionHandler} color='inherit' className={className}>
-      {text}
-    </Button>
-  )
-}
-
-ActionButton.propTypes = {
-  actionHandler: PropTypes.func.isRequired,
-  text: PropTypes.string.isRequired,
-  className: PropTypes.string
-}
