+++ client/src/components/admin/AllApplicantsAdminView.js
