+++ client/src/components/comments/ApplicationComment.js
@@ -34,8 +34,8 @@ class ApplicationComment extends Component {
         <div className='comment__attachments'>
           {attachments.map((attachment, index) => {
             return (
-              <div className='comment__attachment'>
-                <Tooltip key={index+'tooltip'} title={attachment.path.substring(57)}>
+              <div className='comment__attachment' key={index}>
+                <Tooltip key={index} title={attachment.path.substring(57)}>
                   <Chip
                     key={index}
                     label={this.truncateString(attachment.path.substring(57), 10)}
