+++ client/src/components/header/MobileMenu.js
@@ -6,8 +6,8 @@ import PropTypes from 'prop-types'
 
 import Menu from '@material-ui/icons/Menu'
 import ExitToApp from '@material-ui/icons/ExitToApp'
-import * as actions from '../redux/actions/actions'
-import * as selectors from '../redux/selectors/selectors'
+import * as actions from '../../redux/actions/actions'
+import * as selectors from '../../redux/selectors/selectors'
 
 export class MobileMenu extends Component {
   constructor(props) {
