+++ client/src/components/jobPosting/redirect/ApplicationSuccess.js
