+++ client/src/components/jobPostings/LogoutSnackbar.js
